move some capabilities to capabilities module
